Fix atoi32 for large numbers on 64 bit OSes.
Previously it would truncate the result instead of returning an error.

Because int is 64 bits on 64 bit operating systems, strconv.Atoi will
return a 64 bit integer, which gets truncated to 32 bits without
checking to see if it fits. strconv.ParseInt takes a bit count and
verifies that the number fits before returning it.