Fix panic in recent replays
New replays have a CDOTAPlayer.m_hVehicle field that was being treated
as an unsigned varint, but isn’t actually one. Manta was reading 8 bytes
and continuing to get continue bits, causing it to panic. This patch
decodes unsigned varints as 32-bytes by default, with specific
exceptions carved out for known 64-bit varints. This is a safer
approach, thought not a real fix for properly reading the data type.